PRA-62 Fix type of points + comments
