+++ packages/client/src/pages/LeaderboardPage/LeaderboardRow.tsx
@@ -1,7 +1,6 @@
-import defaultAvatar from 'assets/icons/default_avatar.svg';
 import styles from './leaderboard.module.scss';
 import { LeaderboardInfo } from './types';
-import { Avatar } from '../../components';
+import { Avatar } from 'components';
 
 interface LeaderboardRowProps extends Partial<LeaderboardInfo> {
   place: number;
@@ -17,10 +16,7 @@ export function LeaderboardRow({
   return (
     <li className={styles.leaderboardRow}>
       <p>{place}</p>
-      <Avatar
-        image={userAvatar ? userAvatar : defaultAvatar}
-        className={styles.avatar}
-      />
+      <Avatar image={userAvatar ?? userAvatar} className={styles.avatar} />
       <p>{userNickname}</p>
       <p>{points}</p>
       <p>{time}</p>
