--- packages/client/src/pages/LeaderboardPage/data.ts
@@ -1,22 +1,0 @@
-export const leaderboardFields = [
-  {
-    id: 'place',
-    title: 'Рейтинг',
-  },
-  {
-    id: 'avatar',
-    title: '',
-  },
-  {
-    id: 'username',
-    title: 'Никнейм',
-  },
-  {
-    id: 'score',
-    title: 'Очки',
-  },
-  {
-    id: 'time',
-    title: 'Время, c.',
-  },
-];
